Trailers partially implemented
A lot of work left to do. I need an onStart or onResume for the details
activity: currently trying to navigate back to it from the reviews page
crashes. The JSON data I'm getting is weird - it seems like it's getting
cut off. I'll likely need to troubleshoot the adapter and layout after I
can get the data right in the first place.