Reviews* still partially implemented.
Added some onSaveINstanceState stuff to the deatils activity - it looks
like it works fine when going back to details with the back button, but
not the navigate upwards taskbar button: it seems like the instance
state gets saved fine navigating to the reviews, but is null when coming
back only using the taskbar back.

I refactored the reviews activity so that it is responsible for fetching
the reviews instead of the details activity. I don't know why i didn't
do that before. Nevertheless it still doesn't work.